requirements.txt
Django==3.2.5
djangorestframework>=3.11.0,<3.12.0
black==21.6b0
psycopg2==2.9.1